Major version bump to 2.0.0
My initial work here attempted to apply fairly strong typing to the
Watson Discovery Service API. This turns out to be overly brittle. The
WDS API continues to evolve. The service developers consider _adding_
a new field to be a non-breaking change, but it would (typically) be a
breaking change for version 1 of this library.

Version 2 will be far looser with typing. This pushes more work to the
client code using this library, but does make is possible to remain
functional as the WDS API evolves.